Changed cluster join code to run if not joined.
Currently if initial server is not up while starting other servers or agent, the
cluster join will be failed, and unless there is any (new) service definition,
that node will not be joining to the consul cluster, until there is a service
restart as cluster join code in the is only called after service restart.
This patch will fix this issue.