sysv & debian init updates to kill or leave
This fixes #85. The 'stop' action in init scripts for sysv and Debian
will only 'leave' the cluster if acting as an agent. When running as a
server, as determined by a call to `consul info`, kill the process
instead.

Both updates also enforce the use of a PID file located in a
/var/run/consul directory, writeable by the consul::user configured in
Puppet.