Removed is_systemd, going with sane default
As per AJ and SolarKennedy, we do provide $init_style as a parameter available in the main class.
We should provide sane defaults for the most common cases but not the far edge cases, if the current ability for the user to modify it is already there.
Changed Ubuntu 15.04 and greater to default systemd.