Added cleaning of old staged versions of base consul
Also set consul_web_ui back to what it was, it seems it does support version
just doesn't clean up after itself, it's not big so it's not in my plans to fix
unless requested.
From watching consul start/stop it does seem like it would prefer we left the group
rather than just quit the service...  Should we fix this?