Add a check for $service_ensure in reload_service
When pre-provisioning a node with a tool like Packer (https://packer.io)
it's possible that we're only installing using Puppet, not attempting to
run a live configuration.  "reload_service" didn't respect the setting
of $service_ensure, so it would attempt to reload the consul config even
if the service was supposed to be stopped.

In a pre-provisioning case, the Consul config may be populated with
place-holders for values like keys or servers that are expected to be
replaced at launch time (example: AWS EC2 user-data).

The "reload_service" class is a notification dependency for many things,
so disabling it at each point is difficult and error-prone.  This commit
has "reload_service" ensure that the value of $service_ensure is not
'stopped', and only performs the reload if it's appropriate.

Otherwise, the method is a no-op.