Rework sysv stop script to fix issues
The sysv stop script had been set to issue a SIGKILL (-9) to ensure that
a server would not perform a leave operation.  That's excessive, and the
logic in that section had a race condition that could result in a client
unpredictably ending in either a "left" or "failed" state.

Reworked this section to issue a SIGINT to clients (which initiates a
leave by default) and then check for success.  If the leave operation
fails or if the agent is acting as a server, then the standard killproc
is issued, which tries a SIGTERM, followed by a SIGKILL if the TERM is
not effective.

Also fixed a minor issue where the stop return status would often be
displayed in the prompt line.  The start status is eaten by Consul's
default stdout logging, so there's no fixing that.  It simply ends up in
the log file.