Revert removal of $PID_FILE
The $PID_FILE was being removed by the stop command.  Before the race
condition for the INT/leave condition was fixed, the "rm" could remove
the file before consul itself tried to, which made consul unhappy in the
logs.

Re-add the rm -f $PID_FILE to ensure its removed in a drastic SIGKILL
case.  The waits on INT or TERM will ensure that the rm only happens
when it's needed.