mitigate .to_json segfaults on Ruby 1.8.7
Regularly we had processes segfaulting on line 13 of
consul_sorted_json.rb and we then even had ghost processes (not
tracked by passenger anymore) consuming 100% of memory doing nothing
when attaching with strace.

This might be a problem of either Ruby 1.8.7 itself (similar to
https://groups.google.com/forum/#!msg/puppet-users/ag1rAF7o_Uw/1hjtABrmzwgJ)
or the used c json library. Looking at what we are doing on line 13,
we figured out, that we could reduce the calls to .to_json and
hence might be able to mitigate the segfaults.

Only changing line 13 helped but moved the segfaults to line 10,
looking at that line, we can actually avoid further .to_json calls.

Running this patch on 2 puppet masters for 8 hours didn't show any
segfaults, nor high running cpu processes, which usually started to
appear quite quickly.