Remove puppet error when ACLs cannot be retrieved
Currently if consul does not come up or there is some issue electing a
leader, the consul_acl will throw a hard puppet error. This causes the
puppet run to immediately fail. This commit changes that puppet error to
a warning. The unavailability of the consul API should not cause a
puppet failure.