Cleanup Systemd unit file
Remove references to basic.target: Systemd automatically adds
dependencies of the types Requires= and After= for this target unit to
all services, so KISS and DRY.

Remove KillMode= option: Consul itself runs entirely in one process, so
the default of killing all processes in the control group would not make
a difference. If Consul does spawn processes, e.g. for health checks,
the previous setting would have left them running behind when the
service is stopped.

Remove RestartSec= option: There is no reason to override the default of
100ms.