Revert to default Systemd KillSignal of SIGTERM
Since Consul 0.7, client-mode agents are shutting down gracefully by
default, so no need to send them a SIGINT specially. The default for
server-mode agents is still not to leave gracefully on SIGTERM, but it
was left this way for a reason. To override this, it is better to set
the leave_on_terminate option, rather than changing the KillSignal.