Shared consul code and parent provider for consul_acl
This commit lays the structure for code re-use between providers.
The consul_acl provider is the first to make use of this structure.

The parent providers handles most of the flow used in providers.
The specific providers only have to implement a few methods for crud operations and caching.

The consulclient library is used to talk to consul. It takes care of:
- Authentication
- Retry's
- Decoding of the KV stor
- Some error checking.

Spec_Helper_provider implements a set of tests that can be used by
the providers if the api is similar enough.