Fix validate command for config files in consul 1.0.0
The consul validate command has a problem with using files which do not end in .json or .hcl which is the fact when using validate_cmd which uses a temporary file and puppet adds a tmp-specific string. This solution is kind of an ugly workaround which produces a tmp link file with the correct name and leaves it in /tmp. After the check this link will become invalid and there are no security issues.