Support space-separated list in addresses
"In Consul 1.0 and later [addresses from config_hash] can be set to a space-separated list of addresses to bind to [...]" https://www.consul.io/docs/agent/options.html#addresses

Using a space-separated list in addresses for 'http' is currently not supported because it would break config in init scripts and other places. Limiting this to the first one mentioned fixes this.