Add parameter data types. Move static defaults to init.pp.
This commit adds parameter data types to the main classes and defines.

It also moves any static parameter defaults from params.pp to init.pp.
It's generally better to have static defaults in init.pp and leave
params.pp for any defaults that are calculated. This makes it easier to
see parameter default values when using the console in Puppet Enterprise
as well as reduces the amount of page flipping someone has to do when
casually browsing the code.