Make stringification a parameter of the sorted_json helper functions
Move the 'quoted'ness of child values to be a function parameter instead of a
local variable to avoid an issue where keys processed after a
node_meta|meta|tags key were inadvertently quoted. This should allow it to
track down the recursion as intended.

The escape hatch for SRV weights should no longer be needed.

The base simple_generate() function has been simplified to jsonify
either a string interpolation of int, float, and boolean or the actual value
depending on the quoting request of the parent.