PDK update, move CI to GH Actions
This includes many changes that are a byproduct of updating to the
latest version of the PDK. All unit tests pass at this stage but
several RuboCop tests still fail.

Tht GH Action was copied from a Vox Pupuli module.
Their module testing is pretty slick and utilizes Beaker just as this
one has traditionally done so it seems like a good place to start. A
couple of lines were commented out as the Gemfile from the PDK doesn't
line up exactly with the one used by most Vox Pupuli modules.

RuboCop is disabled for the moment as it is known that those tests do
not yet pass.